fix(recent-tabs): match station icons and truncate labels
Pre-commit hook ran. Total eslint: 0, total circular: 1
