+++ docs/frontend-ui-audit-2026-09-05/RecentTabsMenuSection.md
@@ -1,9 +1,12 @@
 # RecentTabsMenuSection UI audit
 
-| Line                                 | Element                     | Verdict          | Reason                                                                                                                                | Suggested change |
-| ------------------------------------ | --------------------------- | ---------------- | ------------------------------------------------------------------------------------------------------------------------------------- | ---------------- |
-| `RecentTabsMenuSection/index.tsx:30` | Section separator and label | keep with reason | Reuses shared dropdown separator and section-label tokens, keeping spacing, borders, typography, and themes aligned with other menus. | None.            |
-| `RecentTabsMenuSection/index.tsx:36` | Recent-tab action rows      | keep with reason | The shared scaffold owns native menu-item buttons and applies the canonical `menuActionItem` token for keyboard and button semantics. | None.            |
-| `RecentTabsMenuSection/index.tsx:31` | Labeled menu group          | keep with reason | `role="group"` and `aria-labelledby` expose the section heading while each recent row remains an actual button.                       | None.            |
+| Line                                                                          | Element                     | Verdict          | Reason                                                                                                                                                              | Suggested change |
+| ----------------------------------------------------------------------------- | --------------------------- | ---------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------------- |
+| `RecentTabsMenuSection/index.tsx:30`                                          | Section separator and label | keep with reason | Reuses shared dropdown separator and section-label tokens, keeping spacing, borders, typography, and themes aligned with other menus.                               | None.            |
+| `RecentTabsMenuSection/index.tsx:36`                                          | Recent-tab action rows      | keep with reason | The shared scaffold owns native menu-item buttons and applies the canonical `menuActionItem` token for keyboard and button semantics.                               | None.            |
+| `RecentTabsMenuSection/index.tsx:31`                                          | Labeled menu group          | keep with reason | `role="group"` and `aria-labelledby` expose the section heading while each recent row remains an actual button.                                                     | None.            |
+| `src/modules/WorkStation/shared/TabBar/components/WorkstationTabIcon.tsx:119` | Canonical tab-icon renderer | keep with reason | One renderer must select among integration, session, favicon, file-type, and named-glyph sources; no design-system primitive can express that variant set.          | None.            |
+| `src/modules/WorkStation/AppShell/TabBarPlusMenu/TabBarPlusMenu.tsx:84`       | Recent-tab data mapping     | keep with reason | The Workstation menu passes its tab's canonical icon to the shared, lightweight menu surface, avoiding duplicated selection logic or a ChatPanel bundle dependency. | None.            |
+| `src/modules/WorkStation/shared/TabBar/components/SortableTab/index.tsx:228`  | Station tab icon slot       | keep with reason | The existing `TabPillSurface` remains the design-system shell; it consumes the same canonical icon renderer as the recent menu.                                     | None.            |
 
-Verdict totals: **0 fix**, **3 keep with reason**, **0 abstract**.
+Verdict totals: **0 fix**, **6 keep with reason**, **0 abstract**.
